deps(graphql_flutter): use hive_ce instead of hive
The `hive` package is no longer maintained and has been replaced by `hive_ce`, a community-maintained fork.

This change updates `graphql_flutter` to use `hive_ce` and also temporarily disables publishing of the `graphql` package due to a dependency issue with the pre-release version of `graphql` that `graphql_flutter` depends on.